+++ src/main/java/com/spring/attandance/service/GroupService.java
@@ -9,6 +9,7 @@ import com.spring.attandance.domain.Member;
 import com.spring.attandance.repository.GroupMemberRepository;
 import com.spring.attandance.repository.MemberRepository;
 import com.spring.attandance.repository.GroupRepository;
+import com.spring.attandance.service.validator.GroupValidator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
