+++ src/main/java/com/spring/attandance/service/validator/GroupValidator.java
@@ -1,4 +1,4 @@
-package com.spring.attandance.service;
+package com.spring.attandance.service.validator;
 
 public interface GroupValidator {
     void exceedLimitGroupCount(Long memberId);
