+++ src/main/java/com/spring/attandance/service/validator/MemberValidatorImpl.java
@@ -1,6 +1,5 @@
-package com.spring.attandance.service;
+package com.spring.attandance.service.validator;
 
-import com.spring.attandance.domain.Group;
 import com.spring.attandance.domain.Member;
 import com.spring.attandance.domain.Study;
 import com.spring.attandance.domain.cond.MemberSearchCondition;
